# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Transport Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface StaffMember {
  id: number;
  name: string;
  phone: string;
  email: string;
  role: string;
  isActive: boolean;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  staff: StaffMember;
}

export interface Hub {
  id: number;
  hubName: string;
  hubCode: string;
  address: string;
  contactNumber: string;
  isActive: boolean;
  createdAt: string;
}

export interface HubInput {
  hubName: string;
  hubCode: string;
  address: string;
  contactNumber: string;
}

export interface HubUpdate {
  hubName?: string;
  hubCode?: string;
  address?: string;
  contactNumber?: string;
}

export interface StaffUpdate {
  name?: string;
  phone?: string;
  email?: string;
  role?: string;
}

export interface StaffInput {
  name: string;
  phone: string;
  email: string;
  password: string;
  role: string;
}

export interface Parcel {
  id: number;
  awbNumber: string;
  senderName: string;
  senderPhone: string;
  /** @nullable */
  senderEmail?: string | null;
  senderAddress?: string;
  receiverName: string;
  receiverPhone: string;
  /** @nullable */
  receiverEmail?: string | null;
  receiverAddress?: string;
  numBoxes: number;
  weightKg: number;
  itemId: number;
  /** @nullable */
  itemName?: string | null;
  charges: number;
  /** @nullable */
  remarks?: string | null;
  destinationHubId: number;
  /** @nullable */
  destinationHubName?: string | null;
  /** @nullable */
  destinationHubCode?: string | null;
  currentStatus: string;
  /** @nullable */
  bookedBy?: number | null;
  /** @nullable */
  bookedByName?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ParcelStatusHistory {
  id: number;
  parcelId: number;
  status: string;
  /** @nullable */
  hubId?: number | null;
  /** @nullable */
  hubName?: string | null;
  /** @nullable */
  updatedBy?: number | null;
  /** @nullable */
  updatedByName?: string | null;
  /** @nullable */
  notes?: string | null;
  timestamp: string;
}

export interface ParcelDetail {
  id: number;
  awbNumber: string;
  senderName: string;
  senderPhone: string;
  /** @nullable */
  senderEmail?: string | null;
  senderAddress?: string;
  receiverName: string;
  receiverPhone: string;
  /** @nullable */
  receiverEmail?: string | null;
  receiverAddress?: string;
  numBoxes: number;
  weightKg: number;
  itemId: number;
  /** @nullable */
  itemName?: string | null;
  charges: number;
  /** @nullable */
  remarks?: string | null;
  destinationHubId: number;
  /** @nullable */
  destinationHubName?: string | null;
  /** @nullable */
  destinationHubCode?: string | null;
  currentStatus: string;
  /** @nullable */
  bookedBy?: number | null;
  /** @nullable */
  bookedByName?: string | null;
  createdAt: string;
  updatedAt: string;
  history: ParcelStatusHistory[];
}

export interface ParcelListResponse {
  parcels: Parcel[];
  total: number;
  page: number;
  limit: number;
}

export interface ParcelInput {
  senderName: string;
  senderPhone: string;
  senderEmail?: string;
  senderAddress: string;
  receiverName: string;
  receiverPhone: string;
  receiverEmail?: string;
  receiverAddress: string;
  numBoxes: number;
  weightKg: number;
  itemId: number;
  charges: number;
  remarks?: string;
  destinationHubId: number;
}

export interface ParcelUpdate {
  senderName?: string;
  senderPhone?: string;
  senderEmail?: string;
  senderAddress?: string;
  receiverName?: string;
  receiverPhone?: string;
  receiverEmail?: string;
  receiverAddress?: string;
  numBoxes?: number;
  weightKg?: number;
  itemId?: number;
  charges?: number;
  remarks?: string;
}

export interface ParcelStatusUpdate {
  status: string;
  notes?: string;
  hubId?: number;
}

export interface ScanInput {
  awbNumber: string;
  action: string;
  hubId?: number;
  notes?: string;
}

export interface ScanResult {
  parcel: ParcelDetail;
  previousStatus: string;
  newStatus: string;
  actionTaken: string;
}

export interface DashboardStats {
  todayBookings: number;
  incomingParcels: number;
  outgoingParcels: number;
  readyForPickup: number;
  deliveredToday: number;
  openComplaints: number;
}

export interface HubStats {
  hubId: number;
  hubName: string;
  hubCode: string;
  todayBookings: number;
  inTransit: number;
  delivered: number;
  pending: number;
  complaints: number;
}

export interface Complaint {
  id: number;
  complaintNumber: string;
  parcelId: number;
  /** @nullable */
  awbNumber?: string | null;
  raisedBy: number;
  /** @nullable */
  raisedByName?: string | null;
  complaintType: string;
  description: string;
  images?: string[];
  status: string;
  /** @nullable */
  assignedTo?: number | null;
  /** @nullable */
  assignedToName?: string | null;
  /** @nullable */
  resolutionNotes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ComplaintInput {
  parcelId: number;
  complaintType: string;
  description: string;
  images?: string[];
}

export interface ComplaintUpdate {
  status?: string;
  /** @nullable */
  assignedTo?: number | null;
  resolutionNotes?: string;
}

export interface DailyReport {
  date: string;
  totalBooked: number;
  totalDelivered: number;
  totalPending: number;
  revenueToday: number;
}

export interface HubReport {
  hubId: number;
  hubName: string;
  hubCode: string;
  bookings: number;
  deliveries: number;
  pending: number;
  revenue: number;
  complaints: number;
}

export interface MonthlyReport {
  month: number;
  year: number;
  label?: string;
  totalParcels: number;
  delivered: number;
  deliveryPercentage: number;
  complaintRate: number;
  revenue?: number;
}

export interface ExportResponse {
  url: string;
}

/**
 * @nullable
 */
export type AuditLogOldValue = { [key: string]: unknown } | null;

/**
 * @nullable
 */
export type AuditLogNewValue = { [key: string]: unknown } | null;

export interface AuditLog {
  id: number;
  action: string;
  entityType: string;
  entityId: string;
  /** @nullable */
  oldValue?: AuditLogOldValue;
  /** @nullable */
  newValue?: AuditLogNewValue;
  performedBy: number;
  /** @nullable */
  performedByName?: string | null;
  /** @nullable */
  hubId?: number | null;
  /** @nullable */
  hubName?: string | null;
  timestamp: string;
  /** @nullable */
  description?: string | null;
}

export interface AuditLogListResponse {
  logs: AuditLog[];
  total: number;
  page: number;
  limit: number;
}

export interface Item {
  id: number;
  name: string;
  defaultPrice: number;
  createdAt: string;
}

export interface ItemInput {
  name: string;
  defaultPrice: number;
}

export type ListStaffParams = {
hubId?: number;
role?: string;
};

export type ListParcelsParams = {
hubId?: number;
status?: string;
page?: number;
limit?: number;
dateFrom?: string;
dateTo?: string;
};

export type GetDashboardStatsParams = {
hubId?: number;
};

export type GetRecentParcelsParams = {
hubId?: number;
limit?: number;
};

export type ListComplaintsParams = {
hubId?: number;
status?: string;
parcelId?: number;
};

export type GetDailyReportParams = {
date?: string;
hubId?: number;
};

export type GetHubWiseReportParams = {
dateFrom?: string;
dateTo?: string;
};

export type GetMonthlyReportParams = {
months?: number;
};

export type ExportReportParams = {
type: string;
dateFrom?: string;
dateTo?: string;
hubId?: number;
};

export type SearchParcelsParams = {
q?: string;
phone?: string;
dateFrom?: string;
dateTo?: string;
};

export type ListAuditLogsParams = {
staffId?: number;
hubId?: number;
action?: string;
dateFrom?: string;
dateTo?: string;
page?: number;
limit?: number;
};