# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Transport Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /parcels:
    get:
      operationId: listParcels
      tags: [parcels]
      summary: List parcels
      parameters:
        - name: hubId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of parcels
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParcelListResponse"

    post:
      operationId: createParcel
      tags: [parcels]
      summary: Create a parcel booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParcelInput"
      responses:
        "201":
          description: Parcel created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Parcel"
  /parcels/{parcelId}:
    get:
      operationId: getParcel
      tags: [parcels]
      summary: Get parcel details
      parameters:
        - name: parcelId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Parcel details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParcelDetail"

    patch:
      operationId: updateParcel
      tags: [parcels]
      summary: Update parcel details
      parameters:
        - name: parcelId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParcelUpdate"
      responses:
        "200":
          description: Parcel updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Parcel"
...
  /parcels/awb/{awbNumber}:
    get:
      operationId: getParcelByAwb
      tags: [parcels]
      summary: Get parcel by AWB number
      parameters:
        - name: awbNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Parcel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParcelDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # SCANNING
...
  /complaints:
    get:
      operationId: listComplaints
      tags: [complaints]
      summary: List complaints
      parameters:
        - name: hubId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: parcelId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of complaints
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Complaint"

    post:
      operationId: createComplaint
      tags: [complaints]
      summary: Raise a complaint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplaintInput"
      responses:
        "201":
          description: Complaint created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
  /complaints/{complaintId}:
    get:
      operationId: getComplaint
      tags: [complaints]
      summary: Get complaint details
      parameters:
        - name: complaintId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Complaint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"

    patch:
      operationId: updateComplaint
      tags: [complaints]
      summary: Update complaint (assign, resolve, close)
      parameters:
        - name: complaintId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplaintUpdate"
      responses:
        "200":
          description: Complaint updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"

  # REPORTS
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, staff]
      properties:
        token:
          type: string
        staff:
          $ref: "#/components/schemas/StaffMember"
...
    StaffMember:
      type: object
      required: [id, name, phone, email, role, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        role:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
    Parcel:
      type: object
      required: [id, awbNumber, senderName, senderPhone, receiverName, receiverPhone, numBoxes, weightKg, itemId, charges, destinationHubId, currentStatus, createdAt, updatedAt]
      properties:
        id:
          type: integer
        awbNumber:
          type: string
        senderName:
          type: string
        senderPhone:
          type: string
        senderEmail:
          type: ["string", "null"]
        senderAddress:
          type: string
        receiverName:
          type: string
        receiverPhone:
          type: string
        receiverEmail:
          type: ["string", "null"]
        receiverAddress:
          type: string
        numBoxes:
          type: integer
        weightKg:
          type: number
        itemId:
          type: integer
        itemName:
          type: ["string", "null"]
        charges:
          type: number
        remarks:
          type: ["string", "null"]
        destinationHubId:
          type: integer
        destinationHubName:
          type: ["string", "null"]
        destinationHubCode:
          type: ["string", "null"]
        currentStatus:
          type: string
        bookedBy:
          type: ["integer", "null"]
        bookedByName:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ParcelDetail:
      type: object
      required: [id, awbNumber, senderName, senderPhone, receiverName, receiverPhone, numBoxes, weightKg, itemId, charges, destinationHubId, currentStatus, createdAt, updatedAt, history]
      properties:
        id:
          type: integer
        awbNumber:
          type: string
        senderName:
          type: string
        senderPhone:
          type: string
        senderEmail:
          type: ["string", "null"]
        senderAddress:
          type: string
        receiverName:
          type: string
        receiverPhone:
          type: string
        receiverEmail:
          type: ["string", "null"]
        receiverAddress:
          type: string
        numBoxes:
          type: integer
        weightKg:
          type: number
        itemId:
          type: integer
        itemName:
          type: ["string", "null"]
        charges:
          type: number
        remarks:
          type: ["string", "null"]
        destinationHubId:
          type: integer
        destinationHubName:
          type: ["string", "null"]
        destinationHubCode:
          type: ["string", "null"]
        currentStatus:
          type: string
        bookedBy:
          type: ["integer", "null"]
        bookedByName:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        history:
          type: array
          items:
            $ref: "#/components/schemas/ParcelStatusHistory"
    ParcelListResponse:
      type: object
      required: [parcels, total, page, limit]
      properties:
        parcels:
          type: array
          items:
            $ref: "#/components/schemas/Parcel"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    ParcelInput:
      type: object
      required: [senderName, senderPhone, senderAddress, receiverName, receiverPhone, receiverAddress, numBoxes, weightKg, itemId, charges, destinationHubId]
      properties:
        senderName:
          type: string
        senderPhone:
          type: string
        senderEmail:
          type: string
        senderAddress:
          type: string
        receiverName:
          type: string
        receiverPhone:
          type: string
        receiverEmail:
          type: string
        receiverAddress:
          type: string
        numBoxes:
          type: integer
        weightKg:
          type: number
        itemId:
          type: integer
        charges:
          type: number
        remarks:
          type: string
        destinationHubId:
          type: integer
    ParcelUpdate:
      type: object
      properties:
        senderName:
          type: string
        senderPhone:
          type: string
        senderEmail:
          type: string
        senderAddress:
          type: string
        receiverName:
          type: string
        receiverPhone:
          type: string
        receiverEmail:
          type: string
        receiverAddress:
          type: string
        numBoxes:
          type: integer
        weightKg:
          type: number
        itemId:
          type: integer
        charges:
          type: number
        remarks:
          type: string
...
    ParcelStatusHistory:
      type: object
      required: [id, parcelId, status, timestamp]
      properties:
        id:
          type: integer
        parcelId:
          type: integer
        status:
          type: string
        hubId:
          type: ["integer", "null"]
        hubName:
          type: ["string", "null"]
        updatedBy:
          type: ["integer", "null"]
        updatedByName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        timestamp:
          type: string
...
    Complaint:
      type: object
      required: [id, complaintNumber, parcelId, raisedBy, complaintType, description, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        complaintNumber:
          type: string
        parcelId:
          type: integer
        awbNumber:
          type: ["string", "null"]
        raisedBy:
          type: integer
        raisedByName:
          type: ["string", "null"]
        complaintType:
          type: string
        description:
          type: string
        images:
          type: array
          items:
            type: string
        status:
          type: string
        assignedTo:
          type: ["integer", "null"]
        assignedToName:
          type: ["string", "null"]
        resolutionNotes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ComplaintInput:
      type: object
      required: [parcelId, complaintType, description]
      properties:
        parcelId:
          type: integer
        complaintType:
          type: string
        description:
          type: string
        images:
          type: array
          items:
            type: string
    ComplaintUpdate:
      type: object
      properties:
        status:
          type: string
        assignedTo:
          type: ["integer", "null"]
        resolutionNotes:
          type: string
...