# lib/api-zod/src/generated/types/parcelDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Transport Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { ParcelStatusHistory } from './parcelStatusHistory';

export interface ParcelDetail {
  id: number;
  awbNumber: string;
  senderName: string;
  senderPhone: string;
  /** @nullable */
  senderEmail?: string | null;
  senderAddress?: string;
  receiverName: string;
  receiverPhone: string;
  /** @nullable */
  receiverEmail?: string | null;
  receiverAddress?: string;
  numBoxes: number;
  weightKg: number;
  itemId: number;
  /** @nullable */
  itemName?: string | null;
  charges: number;
  /** @nullable */
  remarks?: string | null;
  destinationHubId: number;
  /** @nullable */
  destinationHubName?: string | null;
  /** @nullable */
  destinationHubCode?: string | null;
  currentStatus: string;
  /** @nullable */
  bookedBy?: number | null;
  /** @nullable */
  bookedByName?: string | null;
  createdAt: string;
  updatedAt: string;
  history: ParcelStatusHistory[];
}