# lib/api-zod/src/generated/types/parcelInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Transport Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface ParcelInput {
  senderName: string;
  senderPhone: string;
  senderEmail?: string;
  senderAddress: string;
  receiverName: string;
  receiverPhone: string;
  receiverEmail?: string;
  receiverAddress: string;
  numBoxes: number;
  weightKg: number;
  itemId: number;
  charges: number;
  remarks?: string;
  destinationHubId: number;
}